requirements.txt
altair==5.2.0
extra-streamlit-components==0.1.60
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
millify==0.1.1
monotonic==1.6
mysql-connector==2.2.9
numpy==1.26.4
pandas==2.1.3
pyodbc==5.0.1
SQLAlchemy==2.0.23
streamlit==1.33.0
streamlit-aggrid==0.3.4.post3
streamlit-authenticator==0.2.3
streamlit-metrics==0.1.0
streamlit-option-menu==0.3.6
streamlit-webrtc==0.47.1